requirements.txt

crewai>=0.60.0
duckduckgo-search>=6.3.5
beautifulsoup4>=4.12.3
requests>=2.32.3
python-dotenv>=1.0.1
typer>=0.12.5
rich>=13.9.2

streamlit>=1.38.0
